+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update documentation.
 - Update dependencies.
 
+### Fixed
+
+- Use `as_chunks` in integration tests to avoid `clippy::chunks_exact_to_as_chunks`.
+
 ## [0.1.4] - 2026-08-08
 
 First release of the extension to be published in Zed's [marketplace](https://zed.dev/extensions/word-highlight-lsp).